requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
filelock==3.13.1
fsspec==2023.12.2
ftfy==6.1.3
idna==3.6
jinaai==0.2.9
Jinja2==3.1.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
pillow==10.2.0
pybase64==0.2.0
python-dotenv==1.0.0
regex==2023.12.25
requests==2.31.0
six==1.16.0
sympy==1.12
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.13
